+++ src/services/adapters/__tests__/adapters.test.ts
@@ -115,7 +115,7 @@ describe('adapters', () => {
           ctx({ hasBody: true, hasQuery: true, hasHeaders: true }),
         ),
       ).toBe(
-        'api.url(`/pets/${id}`).query(params).headers(headers).json(body).post().json<Pet>()',
+        'api.url(`/pets/${id}`).query(params ?? {}).headers(headers).json(body).post().json<Pet>()',
       );
     });